Add OperationType to distinguish between transaction and signature operations to block 1271 signatures from session keys

+++ contracts/libraries/SessionLib.sol
@@ -220,7 +220,7 @@ library SessionLib {
       // Disallow self-targeting transactions with session keys as these have the ability to administer
       // the smart account.
       require(address(uint160(transaction.to)) != msg.sender, "Can not target self");
-      require(address(uint256(transaction.to)) != address(this), "Can not manage session keys");
+      require(address(uint160(transaction.to)) != address(this), "Can not manage session keys");
 
       bytes4 selector = bytes4(transaction.data[:4]);
       CallSpec memory callPolicy;
